Move packages around and update docs
