+++ flixelgdx-lwjgl3/src/main/java/org/flixelgdx/backend/lwjgl3/FlixelImGuiDebugOverlay.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.flixelgdx.backend.lwjgl3.debug;
+package org.flixelgdx.backend.lwjgl3;
 
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.Gdx;
